Adding in work from 20200308
I'm going to take this one step at a time. This will be a good exercise
for me, and keep me from going to fast. I want to get tests, structure,
and other things in place from the beginning, but I'll do it in a way
that illustrates concepts for future use.

In other words, the repository will become a teaching tool.