Added in tests.
This required adding in a setup.py, so that I could get

`pytest3 -m pytest`

working.

The blog post about this commit is far more thorough, and walks through
the decisions made in this implementation step. Ultimately, the
important takeaway is that there are now some tests in place, which is
more than I had before this commit.